+++ hqlib/metric/environment/failing_ci_jobs.py
@@ -14,9 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+from typing import Dict
 
 from hqlib import metric_source
 from hqlib.domain import LowerIsBetterMetric
+from hqlib.typing import MetricParameters
 
 
 class FailingCIJobs(LowerIsBetterMetric):
@@ -34,15 +36,15 @@ class FailingCIJobs(LowerIsBetterMetric):
     low_target_value = 2
     metric_source_class = metric_source.Jenkins
 
-    def _parameters(self):
+    def _parameters(self) -> MetricParameters:
         parameters = super()._parameters()
-        parameters['number_of_jobs'] = self._metric_source.number_of_active_jobs()
+        parameters['number_of_jobs'] = str(self._metric_source.number_of_active_jobs())
         return parameters
 
     def value(self):
         url = self._metric_source.failing_jobs_url()
         return -1 if url is None else len(url)
 
-    def url(self):
+    def url(self) -> Dict[str, str]:
         url = self._metric_source.failing_jobs_url()
         return dict() if url is None else url