+++ hqlib/metric/product/automated_regression_test_coverage_metrics.py
@@ -14,10 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Dict
-
 from ... import metric_source
 from ...domain import HigherIsBetterMetric, MetricSourceAgeMetric
+from hqlib.typing import MetricParameters
 
 
 class ARTCoverage(HigherIsBetterMetric):
@@ -47,7 +46,7 @@ class ARTCoverage(HigherIsBetterMetric):
         """ Get the actual coverage measurement from the metric source. """
         raise NotImplementedError  # pragma: nocover
 
-    def _parameters(self) -> Dict[str, str]:
+    def _parameters(self) -> MetricParameters:
         # pylint: disable=protected-access
         parameters = super()._parameters()
         parameters['covered_items'] = self.covered_items