+++ hqlib/metric/product/automated_regression_test_metrics.py
@@ -14,10 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Dict
-
 from ... import metric_source
 from ...domain import LowerIsBetterMetric, MetricSourceAgeMetric
+from hqlib.typing import MetricParameters
 
 
 class FailingRegressionTests(LowerIsBetterMetric):
@@ -39,14 +38,14 @@ class FailingRegressionTests(LowerIsBetterMetric):
             urls = self._get_metric_source_ids()
             return self._metric_source.failed_tests(*urls) + self._metric_source.skipped_tests(*urls)
 
-    def _missing(self):
+    def _missing(self) -> bool:
         urls = self._get_metric_source_ids()
         passed = self._metric_source.passed_tests(*urls)
         failed = self._metric_source.failed_tests(*urls)
         skipped = self._metric_source.skipped_tests(*urls)
         return None in (passed, failed, skipped) or passed < 0 or failed < 0 or skipped < 0
 
-    def _parameters(self) -> Dict[str, str]:
+    def _parameters(self) -> MetricParameters:
         # pylint: disable=protected-access
         parameters = super()._parameters()
         passed_tests = self._metric_source.passed_tests(*self._get_metric_source_ids())