+++ hqlib/metric/product/source_code_metrics.py
@@ -14,10 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Dict
-
 from ..metric_source_mixin import SonarDashboardMetric, SonarViolationsMetric
 from ...domain import LowerPercentageIsBetterMetric
+from hqlib.typing import MetricParameters
 
 
 class CommentedLOC(SonarDashboardMetric, LowerPercentageIsBetterMetric):
@@ -47,7 +46,7 @@ class MethodQualityMetric(SonarViolationsMetric, LowerPercentageIsBetterMetric):
     low_target_value = 5
 
     @classmethod
-    def norm_template_default_values(cls) -> Dict[str, str]:
+    def norm_template_default_values(cls) -> MetricParameters:
         values = super(MethodQualityMetric, cls).norm_template_default_values()
         values['attribute'] = cls.attribute
         return values
@@ -58,7 +57,7 @@ class MethodQualityMetric(SonarViolationsMetric, LowerPercentageIsBetterMetric):
     def _denominator(self):
         return self._metric_source.methods(self._sonar_id())
 
-    def _parameters(self) -> Dict[str, str]:
+    def _parameters(self) -> MetricParameters:
         # pylint: disable=protected-access
         parameters = super()._parameters()
         parameters['attribute'] = self.attribute