+++ hqlib/metric_source/test_report/jenkins_test_report.py
@@ -18,9 +18,11 @@ limitations under the License.
 import ast
 import datetime
 import logging
+from typing import Dict, Optional
 
 from ..abstract import test_report
 from ..url_opener import UrlOpener
+from ...typing import DateTime
 
 
 class JenkinsTestReport(test_report.TestReport):
@@ -28,7 +30,7 @@ class JenkinsTestReport(test_report.TestReport):
     metric_source_name = 'Jenkins testreport'
     needs_metric_source_id = True
 
-    def _passed_tests(self, report_url):
+    def _passed_tests(self, report_url: str) -> int:
         """ Return the number of passed tests. """
         try:
             return self.__test_count(report_url, 'passCount')
@@ -39,25 +41,25 @@ class JenkinsTestReport(test_report.TestReport):
             failed = self._failed_tests(report_url)
             return total - skipped - failed
 
-    def _failed_tests(self, report_url):
+    def _failed_tests(self, report_url: str) -> int:
         """ Return the number of failed tests. """
         return self.__test_count(report_url, 'failCount')
 
-    def _skipped_tests(self, report_url):
+    def _skipped_tests(self, report_url: str) -> int:
         """ Return the number of skipped tests. """
         return self.__test_count(report_url, 'skipCount')
 
-    def __test_count(self, report_url, result_type):
+    def __test_count(self, report_url: str, result_type: str) -> int:
         """ Return the number of tests with the specified result in the test report. """
         json = self.__read_json(self.__join_url(report_url, 'lastCompletedBuild/testReport/api/python'))
         return int(json[result_type]) if json else -1
 
-    def _report_datetime(self, report_url):
+    def _report_datetime(self, report_url: str) -> DateTime:
         """ Return the date and time of the specified report. """
         json = self.__read_json(self.__join_url(report_url, 'lastCompletedBuild/api/python'))
         return datetime.datetime.fromtimestamp(float(json["timestamp"])/1000.) if json else datetime.datetime.min
 
-    def __read_json(self, api_url):
+    def __read_json(self, api_url: str) -> Optional[Dict[str, int]]:
         """ Return the json from the url, or the default when something goes wrong. """
         try:
             contents = self._url_read(api_url)
@@ -70,6 +72,6 @@ class JenkinsTestReport(test_report.TestReport):
             return None
 
     @staticmethod
-    def __join_url(*parts):
+    def __join_url(*parts: str) -> str:
         """ Join the different url parts with forward slashes. """
         return '/'.join([part.strip('/') for part in parts])